Added resilience

+++ src/main/java/ch/cern/todo/controllers/CategoryController.java
@@ -3,6 +3,7 @@ package ch.cern.todo.controllers;
 import ch.cern.todo.controllers.POJO.CategoryPOJO;
 import ch.cern.todo.model.TaskCategory;
 import ch.cern.todo.services.TaskCategoryService;
+import ch.cern.todo.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,6 +20,8 @@ public class CategoryController {
     public ResponseEntity<TaskCategory> newCategory(
             @RequestBody CategoryPOJO category
     ) {
+        if(!Validator.CategoryPOJO(category)) return ResponseEntity.badRequest().build();
+
         var saved = categoryService.save(category);
         return ResponseEntity.ok(saved);
     }
@@ -32,6 +35,8 @@ public class CategoryController {
     public ResponseEntity<String> deleteTask(
             @PathVariable long id
     ) {
+        if(!categoryService.exists(id)) return ResponseEntity.notFound().build();
+
         categoryService.delete(id);
         return ResponseEntity.ok("Deleted");
     }
@@ -40,6 +45,9 @@ public class CategoryController {
     public ResponseEntity<TaskCategory> updateTask(
             @PathVariable long id, @RequestBody CategoryPOJO category
     ) {
+        if(!categoryService.exists(id)) return ResponseEntity.notFound().build();
+        if(!Validator.CategoryPOJO(category)) return ResponseEntity.badRequest().build();
+
         var saved = categoryService.update(id, category);
         return ResponseEntity.of(saved);
     }
