+++ src/main/java/ch/cern/todo/controllers/TaskController.java
@@ -2,7 +2,9 @@ package ch.cern.todo.controllers;
 
 import ch.cern.todo.controllers.POJO.TaskPOJO;
 import ch.cern.todo.model.Task;
+import ch.cern.todo.services.TaskCategoryService;
 import ch.cern.todo.services.TaskService;
+import ch.cern.todo.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +17,8 @@ public class TaskController {
 
     @Autowired
     TaskService taskService;
+    @Autowired
+    TaskCategoryService categoryService;
 
     @GetMapping("/get/{id}")
     public ResponseEntity<Task> getTask(@PathVariable long id) {
@@ -30,6 +34,9 @@ public class TaskController {
     public ResponseEntity<Task> newTask(
             @RequestBody TaskPOJO task
     ) {
+        if(!Validator.taskPOJO(task)) return ResponseEntity.badRequest().build();
+        if(!categoryService.exists(task.getCategoryId())) return ResponseEntity.notFound().build();
+
         var saved = taskService.save(task);
         return ResponseEntity.ok(saved);
     }
@@ -38,6 +45,8 @@ public class TaskController {
     public ResponseEntity<String> deleteTask(
             @PathVariable long id
     ) {
+        if(!taskService.exists(id)) return ResponseEntity.notFound().build();
+
         taskService.delete(id);
         return ResponseEntity.ok("Deleted");
     }
@@ -46,6 +55,13 @@ public class TaskController {
     public ResponseEntity<Task> updateTask(
             @PathVariable long id, @RequestBody TaskPOJO task
     ) {
+        if(!taskService.exists(id)) return ResponseEntity.notFound().build();
+        if(!Validator.taskPOJOAllowNullParameters(task)) return ResponseEntity.badRequest().build();
+        if(
+                task.getCategoryId() != null &&
+                !categoryService.exists(task.getCategoryId())
+        ) return ResponseEntity.notFound().build();
+
         var saved = taskService.update(id, task);
         return ResponseEntity.of(saved);
     }
